Introduced constant variables for algolia search
