+++ packages/website/ts/utils/algolia_search.ts
@@ -1,0 +1,8 @@
+import algoliasearch from 'algoliasearch/lite';
+
+export const searchClient = algoliasearch('39X6WOJZKW', '6acba761a34d99781628c6178af1e16c');
+
+export const searchIndex = {
+    guides: '0x_guides_test',
+    tools: '0x_tools_test',
+};
